Circle CI Working

+++ examples/plot_tikhonov.py
@@ -95,7 +95,7 @@ window = [-100, 100]
 binsize = 20
 
 # Zero pad bins
-n_zero_bins = np.floor((window[1] - window[0]) / binsize / 2)
+n_zero_bins = int(np.floor((window[1] - window[0]) / binsize / 2))
 
 ########################################################
 # Build design matrix
@@ -144,8 +144,8 @@ for fx in fixations_df.index[:1000]:
 
             # Define an image based on the relative locations
             img = np.zeros(shape=(n_shape, n_shape))
-            row = -np.round(probe_row - fix_row) + n_shape / 2 - 1
-            col = np.round(probe_col - fix_col) + n_shape / 2 - 1
+            row = int(-np.round(probe_row - fix_row) + n_shape / 2 - 1)
+            col = int(np.round(probe_col - fix_col) + n_shape / 2 - 1)
             img[row, col] = 1
 
             # Compute projection
